Fix `_get_offset` when run outside of a git repo
Previously, `_get_offset` would raise a `CalledProcessError` when given
an `expanded_dir_path` that was not part of a git repo, because `git`
exits nonzero in this case. This error was caught by the the `except
CalledProcessError` clauses in both `_apply_three_way_patch` and
`_apply_patch_with_rejections`, because their `try` blocks included the
`_get_offset` call. However, the `except` clauses were not updated after
additional steps were added to their `try` blocks, so they erroneously
assumed that the caught errors were raised by running the `git apply`
command. The net result of this is that updates would not be applied
when run outside of git repos.

Resolve this bug in two different ways.

First, because the purpose of `_get_offset` is to find the subdirectory
within a git repository, it's sensible to have it return the empty
string as offset when run outside of a git repository. Add
a `try/except` to the function's body to catch the error when `git
rev-parse` exits nonzero, and if the stderr indicates that the given
`expanded_dir_path` is not a git repo, return the empty string;
otherwise, re-raise the error.

Secondly, to prevent confusion about which command errored in
`_apply_three_way_patch` and `apply_patch_with_rejections`, move
everything besides the `run()` call that runs `git apply ...` outside of
the `try` block, including the `_get_offset` call. This ensures that any
unexpected errors raised by `_get_offset` are not erroneously assumed to
be the result of running `git apply`, so the apply functions will not
mistakenly continue in the belief that they have attempted to apply the
diff.

Finally, correct a typo where the three-way flag `-3` was mistakenly
passed to `git apply` in `_apply_patch_with_rejections`, instead of
`--reject`.